feat: split out fields into own components
