+++ apps/trading/components/deposit-container/fields/index.ts
@@ -1,0 +1,6 @@
+export { FromAddress } from './from-address';
+export { FromChain } from './from-chain';
+export { FromAsset } from './from-asset';
+export { Amount } from './amount';
+export { ToPubKey } from './to-pubkey';
+export { ToAsset } from './to-asset';
